[chore] 컴포넌트를 분리합니다.
